Added setup.py and restructured project.
This setup.py is a draft and some information might still be missing.
A correct email address is missing in any case.

Since python software should be strucuted in packages, I've moved
websocket.py into a sub-folder. This also changes the import path.

Changed import in server.py to the new import path.
