+++ setup.py
@@ -1,0 +1,15 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='websocket_server',
+    version='0.4',
+    packages=find_packages("."),
+    url='https://github.com/Pithikos/python-websockets-server',
+    license='MIT',
+    author='Johan Hanssen Seferidis',
+    author_email='test@example.com',
+    install_requires=[
+    ],
+    description='A simple fully working WebsocketsServer in Python (with no dependencies)',
+    platforms='any',
+)
